Modified Role and Authorities. Added ServerItem access checking feature.

+++ src/main/java/com/ase/serverckecklist/server/service/ServerInfoService.java
@@ -1,5 +1,6 @@
 package com.ase.serverckecklist.server.service;
 
+import com.ase.serverckecklist.file.service.FileService;
 import com.ase.serverckecklist.server.dto.ServerInfoDto;
 import com.ase.serverckecklist.server.entity.ServerInfo;
 import com.ase.serverckecklist.server.repository.CheckListRepository;
@@ -7,9 +8,14 @@ import com.ase.serverckecklist.server.repository.MapRepository;
 import com.ase.serverckecklist.server.repository.MemoRepository;
 import com.ase.serverckecklist.server.repository.ServerInfoRepository;
 import com.ase.serverckecklist.server.vo.ServerInfoVO;
-import com.ase.serverckecklist.file.service.FileService;
+import com.ase.serverckecklist.user.entity.User;
+import com.ase.serverckecklist.user.repository.UserRepository;
+import com.ase.serverckecklist.user.service.UserService;
+import com.ase.serverckecklist.user.vo.UserVO;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
@@ -19,11 +25,13 @@ import java.util.ArrayList;
 @RequiredArgsConstructor
 public class ServerInfoService {
 
+    private final UserRepository userRepository;
     private final ServerInfoRepository serverInfoRepository;
     private final MemoRepository memoRepository;
     private final CheckListRepository checkListRepository;
     private final MapRepository mapRepository;
     private final FileService fileService;
+    private final UserService userService;
 
     // 서버 전체 조회
     public ArrayList<ServerInfoVO> index() {
@@ -34,6 +42,14 @@ public class ServerInfoService {
         ArrayList<ServerInfo> serverList = (ArrayList<ServerInfo>) serverInfoRepository.findAll();
 
         for (ServerInfo serverInfo : serverList) {
+            // 서버 관리자 닉네임
+            User user = userRepository.findByEmail(serverInfo.getManagerId()).orElse(null);
+            String managerNickname = "관리자없음";
+
+            if (user != null) {
+                managerNickname = user.getNickname();
+            }
+
             // 서버 정보, 사진, 각 서버의 메모, 체크리스트, 맵 수를 저장한 VO 생성
             ServerInfoVO vo = ServerInfoVO.builder()
                         .id(serverInfo.getId())
@@ -42,6 +58,7 @@ public class ServerInfoService {
                         .usage(serverInfo.getUsage())
                         .description(serverInfo.getDescription())
                         .managerId(serverInfo.getManagerId())
+                        .managerNickname(managerNickname)
                         .createdDate(serverInfo.getCreatedDate().toString())
                         .modifiedDate(serverInfo.getModifiedDate().toString())
                         .numOfMemo(memoRepository.countByServerId(serverInfo.getId()))
@@ -59,26 +76,37 @@ public class ServerInfoService {
     public ServerInfoVO show(String id) {
         ServerInfo info = serverInfoRepository.findById(id).orElse(null);
 
-        return info != null ?
-                ServerInfoVO.builder()
+        // 빈 값 처리
+        if (info == null) {
+            return null;
+        }
+        
+        // 서버 관리자 닉네임
+        User user = userRepository.findByEmail(info.getManagerId()).orElse(null);
+        String managerNickname = "관리자없음";
+
+        if (user != null) {
+            managerNickname = user.getNickname();
+        }
+
+        return ServerInfoVO.builder()
                     .id(info.getId())
                         .name(info.getName())
                         .photoId(info.getPhotoId())
                         .usage(info.getUsage())
                         .description(info.getDescription())
                         .managerId(info.getManagerId())
+                        .managerNickname(managerNickname)
                         .createdDate(info.getCreatedDate().toString())
                         .modifiedDate(info.getModifiedDate().toString())
-                    .build()
-                : null;
+                    .build();
     }
 
     // 새 서버 추가
-    public ServerInfo create(ServerInfoDto dto) throws IOException {
+    @Transactional
+    public ServerInfo create(HttpServletRequest request, ServerInfoDto dto) throws IOException {
         // 중복id 존재 시 데이터 추가 x
-        if (dto.getId() != null) {
-            return null;
-        }
+        if (dto.getId() != null) return null;
 
         // 파일 id
         String fileId = null;
@@ -90,19 +118,51 @@ public class ServerInfoService {
             }
         }
 
+        // 등록 요청자를 request의 token에서 가져옴
+        UserVO userVO = userService.currentUser(request);
+
+        if (userVO == null) {
+            return null;
+        }
+
+        // body와 token의 등록 요청자(매니저id) 비교
+        if (!userVO.getEmail().equals(dto.getManagerId())) {
+            return null;
+        }
+
         // 파일 id를 넣은 dto를 entity로 변환
         ServerInfo serverInfo = dto.toEntity(fileId);
 
-        return serverInfoRepository.save(serverInfo);
+        // 저장된 서버
+        ServerInfo savedServer = serverInfoRepository.save(serverInfo);
+
+        // 사용자를 해당 서버에 등록
+        userService.joinServer(request, savedServer.getId());
+        // 등록 요청자가 서버 매니저 역할이 없다면 추가
+        userService.promoteServerAdmin(userVO.getEmail());
+
+        return savedServer;
     }
 
     // 서버 수정
-    public ServerInfo update(String id, ServerInfoDto dto) throws IOException {
+    @Transactional
+    public ServerInfo update(HttpServletRequest request, String id, ServerInfoDto dto) throws IOException {
+        // 요청한 사용자 정보 확인
+        UserVO userVO = userService.currentUser(request);
+
+        // 게스트 접근 차단
+        if (userVO == null) return null;
+
         // 수정 대상
         ServerInfo target = serverInfoRepository.findById(id).orElse(null);
 
         // 잘못된 요청 처리
         if (target == null || !id.equals(dto.getId())) {
+            return null;
+        }
+
+        // 서버 관리자가 아니라면 요청 차단
+        if (!userVO.getEmail().equals(target.getManagerId())) {
             return null;
         }
 
@@ -131,7 +191,14 @@ public class ServerInfoService {
 
 
     // 서버 삭제
-    public ServerInfo delete(String id) {
+    @Transactional
+    public ServerInfo delete(HttpServletRequest request, String id) {
+        // 요청한 사용자 정보 확인
+        UserVO userVO = userService.currentUser(request);
+
+        // 게스트 접근 차단
+        if (userVO == null) return null;
+
         ServerInfo target = serverInfoRepository.findById(id).orElse(null);
 
         // 대상이 없으면 잘못된 요청 처리
@@ -139,10 +206,17 @@ public class ServerInfoService {
             return null;
         }
 
+        // 서버 관리자가 아니라면 요청 차단
+        if (!userVO.getEmail().equals(target.getManagerId())) {
+            return null;
+        }
+
         // 서버에 등록된 사진 제거
         if (target.getPhotoId() != null) {
             fileService.deleteFile(target.getPhotoId());
         }
+
+        // ****** 서버 아이템 제거 방법 필요 ******
 
         serverInfoRepository.delete(target);
         return target; // HTTP 응답의 body가 없는 ResponseEntity 생성
